requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-login>=0.6.0
authlib>=1.5.0
requests>=2.25.0
flask-mail>=0.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
gunicorn>=21.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0